+++ packages/mobile/src/overlays/drawer/useDrawerSpacing.ts
@@ -5,24 +5,35 @@ import { MAX_OVER_DRAG } from '@coinbase/cds-common/animation/drawer';
 
 import { useSafeBottomPadding } from '../../hooks/useSafeBottomPadding';
 
-export const useDrawerSpacing = (pin: PinningDirection | undefined = 'bottom') => {
-  const { top } = useSafeAreaInsets();
+export const useDrawerSpacing = (
+  pin: PinningDirection | undefined = 'bottom',
+  disableSafeAreaPaddingBottom: boolean = false,
+) => {
+  const { top: safeTopPadding } = useSafeAreaInsets();
   const safeBottomPadding: number = useSafeBottomPadding();
 
   const safeAreaStyles = useMemo(() => {
     switch (pin) {
       case 'top':
-        return { paddingTop: top + MAX_OVER_DRAG };
+        return { paddingTop: safeTopPadding + MAX_OVER_DRAG };
       case 'left':
-        return { paddingTop: top, paddingLeft: MAX_OVER_DRAG };
+        return { paddingTop: safeTopPadding, paddingLeft: MAX_OVER_DRAG };
       case 'bottom':
-        return { paddingBottom: safeBottomPadding + MAX_OVER_DRAG };
+        return {
+          paddingBottom: disableSafeAreaPaddingBottom
+            ? MAX_OVER_DRAG
+            : safeBottomPadding + MAX_OVER_DRAG,
+        };
       case 'right':
-        return { paddingTop: top, paddingRight: MAX_OVER_DRAG };
+        return { paddingTop: safeTopPadding, paddingRight: MAX_OVER_DRAG };
       default:
-        return { paddingBottom: safeBottomPadding + MAX_OVER_DRAG };
+        return {
+          paddingBottom: disableSafeAreaPaddingBottom
+            ? MAX_OVER_DRAG
+            : safeBottomPadding + MAX_OVER_DRAG,
+        };
     }
-  }, [pin, safeBottomPadding, top]);
+  }, [pin, safeBottomPadding, safeTopPadding, disableSafeAreaPaddingBottom]);
 
   return safeAreaStyles;
 };