+++ packages/mobile/src/overlays/tray/Tray.tsx
@@ -118,6 +118,7 @@ export const Tray = memo(
             flexGrow={1}
             flexShrink={1}
             minHeight={0}
+            overflow="hidden"
             paddingTop={title ? 0 : 2}
             style={contentStyle}
           >
@@ -193,6 +194,9 @@ export const Tray = memo(
   }),
 );
 
+/**
+ * @deprecated Redundant component.
+ */
 export const TrayStickyFooter = ({ children }: { children: ReactNode }) => {
   const { verticalDrawerPercentageOfView, titleHeight } = useContext(TrayContext);
   const { height } = useWindowDimensions();