+++ packages/common/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.55.0 ((3/19/2026, 01:41 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.54.0 ((3/18/2026, 02:27 PM PST))
 
 This is an artificial version bump with no new change.